refactor(user): :recycle: update return types in user use cases to UserAggregate for improved consistency
