+++ src/modules/user/application/queries/use-cases/list-user/list-user.use-case.ts
@@ -1,22 +1,22 @@
 import { Injectable } from '@nestjs/common';
 import { IQueryHandler } from '@nestjs/cqrs';
-import { ListUserModel } from '@user/domain/models/user-list.model';
 
 import { Criteria } from '@common/domain/criteria/criteria';
 import { ListUserQuery } from '@user/application//queries/use-cases/list-user/list-user.query';
-import { ListUserDomainService } from '@user/domain/services/list-user.domain-service';
+import { ListUserAggregate } from '@user/domain/aggregates/list-user.aggregate';
+import { ListUserDomainService } from '@user/domain/domain-services/list-user.domain-service';
 
 @Injectable()
-export class ListUserUseCase implements IQueryHandler<ListUserQuery, ListUserModel> {
+export class ListUserUseCase implements IQueryHandler<ListUserQuery, ListUserAggregate> {
   constructor(private readonly listUserDomainService: ListUserDomainService) {}
 
-  async execute(query: ListUserQuery): Promise<ListUserModel> {
+  async execute(query: ListUserQuery): Promise<ListUserAggregate> {
     const { orderType, orderBy, limit, offset, filtersMap } = query;
 
     const criteria = new Criteria(filtersMap, orderBy, orderType, limit, offset);
 
-    const users = this.listUserDomainService.go(criteria);
+    const listUserAggregate = this.listUserDomainService.go(criteria);
 
-    return users;
+    return listUserAggregate;
   }
 }
